pyproject.toml
[project]
name = "geographie-lernapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "folium>=0.20.0",
    "openai>=1.101.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.48.1",
    "streamlit-authenticator>=0.4.2",
    "streamlit-folium>=0.25.1",
]
